replace int with size_t where needed, in case of 16bit int

+++ src/web.h
@@ -1,14 +1,16 @@
 #ifndef REMOTE_H
 #define REMOTE_H
 
+#include <stdlib.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 struct data_struct {
     char *contents;
-    int allocated;
-    int length;
+    size_t allocated;
+    size_t length;
 };
 
 int get_feed(const char *url, struct data_struct *data);
